refactor: cart 관련 함수 useCart 훅으로 이동
