+++ src/hook/useCart.ts
@@ -1,9 +1,11 @@
 import { useEffect, useState } from 'react';
 import { cartApi } from '../api/cart';
 import { CartItem } from '../types/common';
+import { useToast } from '../component/@common/Toast/context/toastContext';
 
 const useCart = () => {
   const [cartData, setCartData] = useState<CartItem[]>([]);
+  const { openToast } = useToast();
 
   useEffect(() => {
     loadCartData();
@@ -20,7 +22,42 @@ const useCart = () => {
     }
   };
 
-  return { cartData, loadCartData };
+  const addCart = async (productId: number) => {
+    try {
+      await cartApi.addToCart(productId);
+
+      loadCartData();
+      openToast('상품이 장바구니에 추가되었습니다.', true);
+    } catch (error) {
+      openToast('장바구니 담기에 실패했어요...', false);
+    }
+  };
+
+  const removeCart = async (cartId: number) => {
+    try {
+      const cartItem = cartData.filter(
+        (item: CartItem) => item.product.id === cartId
+      );
+
+      if (!cartItem) {
+        console.error('장바구니에서 해당 상품을 찾을 수 없습니다:', cartId);
+        openToast('장바구니에서 상품을 찾을 수 없습니다.', false);
+        return;
+      }
+
+      const targetId = cartItem[0].id;
+
+      await cartApi.removeFromCart(targetId);
+
+      await loadCartData();
+      openToast('상품이 장바구니에서 제거되었습니다.', true);
+    } catch (error) {
+      console.error('장바구니 아이템 삭제 중 오류 발생:', error);
+      openToast('장바구니 빼기에 실패했어요...', false);
+    }
+  };
+
+  return { cartData, loadCartData, addCart, removeCart };
 };
 
 export default useCart;
